requirements.txt
Package             Version
------------------- -------
asgiref             3.8.1
Django              5.1.2
djangorestframework 3.15.2
pip                 24.2
setuptools          75.1.0
sqlparse            0.5.1
wheel               0.44.0
